terraform/airflow/requirements.txt
apache-airflow-providers-google
pandas
pyarrow
gcsfs
